% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaxwell.R
\name{rmaxwell}
\alias{rmaxwell}
\title{Generating samples from Maxwell distribution}
\usage{
rmaxwell(n, sigma)
}
\arguments{
\item{n}{the number of samples}

\item{sigma}{the parameter of Maxwell distribution}
}
\value{
a random sample of size \code{n}
}
\description{
Generating samples from Maxwell distribution
}
\examples{
\dontrun{
rmaxwell(100,2)
}
}
