\name{StatComp20033-package}
\alias{StatComp20033-package}
\alias{StatComp20033}
\docType{package}
\title{
  20033's final project for 'Statistical Computing' course
}
\description{
  20033's final project for 'Statistical Computing' course.
}
\details{
  two-fold is used to leave two cross-validation, just input the parameter dependent variable y and independent variable x.The rmaxwell function is used to generate random numbers that obey the Maxwell distribution. Just enter the number of random numbers to be generated and the parameters of the Maxwell distribution.
}
\author{
Maintainer: RuiPan <machinepan@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
