% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Giao.R
\name{two_fold}
\alias{two_fold}
\title{Use leave-two-out cross validation to compare the models.}
\usage{
two_fold(y, x)
}
\arguments{
\item{y}{the dependent variable data.}

\item{x}{the independent variable data.}
}
\value{
Mean square error of the four models.
}
\description{
Use leave-two-out cross validation to compare the models.
}
\examples{
\dontrun{
two_fold(mtcars$mpg,mtcars$cyl)
}
}
